honor testPathDirs configuration
